requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.0
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.5.30
chardet==4.0.0
cytoolz==0.11.0
Django==3.2.4
djangorestframework==3.12.4
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.17.3
py-solc==3.2.0
pycryptodome==3.10.1
pyrsistent==0.18.0
pytz==2021.1
pywin32==301
requests==2.25.1
rlp==2.0.1
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.1
toolz==0.11.1
typing-extensions==3.10.0.0
urllib3==1.26.6
varint==1.0.2
web3==5.20.0
websockets==8.1
yarl==1.6.3
